moves function outside of src and updates filepaths for searchbar
