+++ frontend/client/functions/index.js
@@ -1,8 +1,34 @@
 const functions = require('firebase-functions');
+const _ = require('lodash');
+const request = require('request-promise');
 
-// // Create and Deploy Your First Cloud Functions
-// // https://firebase.google.com/docs/functions/write-firebase-functions
-//
-// exports.helloWorld = functions.https.onRequest((request, response) => {
-//  response.send("Hello from Firebase!");
-// });
+exports.indexSheltersToElastic = functions.database.ref('/shelters/{resourceId}')
+	.onWrite(event => {
+		let shelterData = event.data.val();
+		let resourceId   = event.params.resourceId;
+
+		console.log('Indexing shelter ', resourceId, shelterData);
+
+		let elasticsearchFields = ['name','keywords','services'];
+		let elasticSearchConfig = functions.config().elasticsearch;
+		let elasticSearchUrl = elasticSearchConfig.url + '/home/shelters/' + resourceId;
+		let elasticSearchMethod = shelterData ? 'POST' : 'DELETE';
+
+		let elasticsearchRequest = {
+			method: elasticSearchMethod,
+			uri: elasticSearchUrl,
+			auth: {
+				username: elasticSearchConfig.username,
+				password: elasticSearchConfig.password,
+			},
+			body: _.pick(shelterData, elasticsearchFields),
+			json: true
+		};
+
+		return request(elasticsearchRequest).then(response => {
+			console.log('Elasticsearch response', response);
+		})
+
+	});
+
+
