+++ frontend/client/src/components/Header.js
@@ -1,6 +1,6 @@
 
 import React from 'react'
-import SearchBar from './Search/searchBar'
+import SearchBar from './searchBar'
 import { Link } from 'react-router-dom'
 import { useStateValue } from '../state/state'
