+++ frontend/client/src/components/searchBar.js
